+++ react/src/lib/components/Layout/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) 2021- Equinor ASA
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import "./Dropdown/dropdown.css";
+import "./Frame/frame.css";
+import "./Header/header.css";
+import "./Label/label.css";
+import "./RadioItems/radioitems.css";
+import "./Selectors/selectors.css";
+import "./Slider/slider.css";
+import "./Tab/tab.css";